Aria: A bit of splitting.
